backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-decorator-include==3.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
importlib-metadata==3.10.0
Markdown==3.3.4
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
zipp==3.4.1
